# doc/api.yaml
openapi: 3.0.3 
servers:
  - url: "http://localhost:3000/v1"
info:
  title: WASAText API
  description: WASAText - Messaging app API spec for WASA course
  version: "1.0.0"
...
paths:
...
  /me/username:
    put:
      tags: ["profile"]
      operationId: setMyUserName
      summary: Change user display name
      description: Change your display name
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload to update user's display name
              properties:
                name:
                  type: string
                  description: User's display name 
                  example: "NewName"
      responses:
        '200':
          description: Name changed successfully
          content:
            application/json:
              schema:
                type: object
                description: Response containing the update user display name
                properties:
                  name: 
                    type: string 
              example:
                name: "NewName"
        '401':
          $ref: '#/components/responses/Unauthorized' 
        '404':
          $ref: '#/components/responses/NotFound' 
        '400':
          $ref: '#/components/responses/BadRequest' 
  /me/photo:
    put:
      tags: ["profile"]
      operationId: setMyPhoto
      summary: Upload or update user profile photo 
      security: 
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: 
              type: object
              description: Payload containing user profile photo 
              properties:
                photo:
                  type: string 
                  format: binary
                  description: Profile image file to upload 
      responses:
        '200':
          description: Photo uploaded
          content:
            application/json:
              schema:
                type: object
                description: Confirmation of photo upload
                properties:
                  message:
                    type: string
                    description: Sucess message
                    example: "Photo uploaded"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound' 
        '400':
          $ref: '#/components/responses/BadRequest' 
...
  /me/conversations:
    get:
      tags: ["conversations"]
      operationId: getMyConversations
      summary: Get all conversations for the current user
      description: Retrieve a list of conversations the current user is part of 
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of conversations returned 
          content:
            application/json:
              schema:
                type: object
                description: List of user's conversations
                properties:
                  items:
                    type: object
                    description: A single conversation summary
                    properties:
                      id:
                        type: integer
                        description: Unique conversation identifier
                        example: 93
                        minimum: 1
                        maximum: 100
                      name:
                        type: string
                        description: Name/title of the conversation
                      lastMessageText:
                        type: string
                        nullable: true
                        description: Snippet of the latest message (if any)
                      lastMessageAt:
                        type: string
                        format: date-time
                        nullable: true
                        description: Timestamp of the latest message (if any)
                      photoUrl:
                        type: string
                        nullable: true
                        description: Conversation or participant photo URL 
                example:
                  - id: 2342
                    name: "Chat with Emanuele"
                    lastMessageText: "Ok, SEE YA"
                    lastMessageAt: "2025-08-15T21:12:03Z"
                    photoUrl: "https://example.com/u/11111.jpg"
        '401': 
          $ref: '#/components/responses/Unauthorized'
        '403': 
          $ref: '#/components/responses/Forbidden'
        '404': 
          $ref: '#/components/responses/NotFound'
        '400': 
          $ref: '#/components/responses/BadRequest'
...
  /users:
    get:
      tags: ["profile"]
      operationId: searchUsers
      summary: Search users by name
      description: Retrieve a list of users whose username contains the given query string.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Substring to match usernames against.
          schema:
            type: string
            description: The text fragment used to search for users.
            example: "juli"
            pattern: '^[A-Za-z0-9_]+$'
            minLength: 1
            maxLength: 32
      responses:
        '200':
          description: List of users matching the search query.
          content:
            application/json:
              schema:
                type: array
                description: List of users found matching the query.
                items:
                  type: object
                  description: Single user search result.
                  properties:
                    id:
                      type: string
                      description: Unique identifier of the user.
                      example: "abc123"
                      pattern: '^[0-9a-zA-Z-]+$'
                      minLength: 3
                      maxLength: 64
                    name:
                      type: string
                      description: Username of the user.
                      example: "juli"
                      pattern: '^[A-Za-z0-9_ ]+$'
                      minLength: 3
                      maxLength: 16
                    photo:
                      type: string
                      nullable: true
                      description: URL of the user profile photo.
                      example: "https://example.com/u/juli.jpg"
...
components:
  responses:
    Unauthorized:
      description: The access token is missing or it's expired
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not authorized"
    BadRequest:
      description: The request was not compliant with the documentation (e.g., missing fields, etc.)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Bad request"
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not found"
    Forbidden:
      description: You are not allowed to perform this operation
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Forbidden operation"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: userId
      description: Unique user identifier returned at login